+++ packages/buefy-next/src/utils/config.js
@@ -35,6 +35,12 @@ let config = {
     defaultTimepickerMobileModal: true,
     defaultNoticeQueue: true,
     defaultInputHasCounter: true,
+    /**
+     * Whether `class`, `style`, and `id` are applied to the root element in
+     * components that are affected by Vue 3 change in fallthgourh beahvior.
+     * See: https://github.com/ntohq/buefy-next/issues/16
+     */
+    defaultCompatFallthrough: true,
     defaultTaginputHasCounter: true,
     defaultUseHtml5Validation: true,
     defaultDropdownMobileModal: true,