+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Buefy Changelog
 
+## buefy-next unreleased
+
+### New features
+
+* `Input` introduces a new prop `compat-fallthrough`, which determines whether `class`, `style`, and `id` attributes are applied to the root `<div>`, or either of `<input>` or `<textarea>` element.
+  If `true`, they are applied to the root `<div>` element, which is compatible with Vue 2.
+  The default value can be controlled by `defaultCompatFallthrough` config option (`true` by default).
+  [#16](https://github.com/ntohq/buefy-next/issues/16)
+
 ## buefy-next
 
 ### Breaking changes